+++ src/models/persistence/DailyMovementCount.ts
@@ -1,3 +1,0 @@
-export class DailyMovementCount {
-  constructor(public deviceID: string, public counter: number, public roomName: string, public date: Date) {}
-}