+++ src/models/persistence/index.ts
@@ -1,6 +1,5 @@
 export * from './BasicRoomInfo';
 export * from './CurrentIlluminationDataPoint';
-export * from './DailyMovementCount';
 export * from './EnergyCalculation';
 export * from './RoomDetailInfo';
 export * from './ShutterCalibration';