+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Remove Daily Movement Count
+
 ## 1.0.64 (2022-09-09)
 
 * (theimo1221) Remove Mongo